Update to use Rust 1.15
Update serde and dependencies that use `#[derive]` from Rust 1.15.
Along with this the build.rs can go away since the code generation
is now automatic.
