+++ .travis.yml
@@ -1,6 +1,5 @@
 language: rust
 rust:
-  - 1.14.0
   - stable
   - beta
   - nightly
