+++ Cargo.toml
@@ -2,16 +2,13 @@
 name = "saytshen"
 version = "0.1.0"
 authors = ["Joseph Wright <rjosephwright@gmail.com>"]
-build = "build.rs"
-
-[build-dependencies]
-serde_codegen = "0.8"
 
 [dependencies]
 clap = "2"
 csv = "0.15"
 yaml-rust = "*"
-serde = "0.8"
-serde_yaml = "0.5"
+serde = "0.9"
+serde_derive = "0.9"
+serde_yaml = "0.6"
 rustc-serialize = "0.3"
 regex = "0.2"
