+++ src/lib.rs
@@ -1,7 +1,10 @@
 extern crate csv;
 extern crate regex;
 extern crate rustc_serialize;
-extern crate serde;
+
+#[macro_use]
+extern crate serde_derive;
+
 extern crate serde_yaml;
 
 #[cfg(test)]
