+++ src/scan.rs
@@ -8,7 +8,26 @@ use csv;
 use regex;
 use serde_yaml;
 
-include!(concat!(env!("OUT_DIR"), "/serde_types.rs"));
+#[derive(Serialize, Deserialize, Debug, Copy, Clone)]
+enum Mode {
+    All,
+    Any,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+struct AuditStep {
+    run: String,
+    expect: Option<String>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+struct Benchmark {
+    section: String,
+    description: String,
+    audit: Vec<AuditStep>,
+    mode: Option<Mode>,
+    skip: Option<String>,
+}
 
 #[derive(Debug, RustcEncodable)]
 struct BenchmarkResult {
